feat(type): update form.settings type to have all options
I got the list of defaults by creating a new form using the API and
inspecting the response as some of the options were not listed on the
docs / had different names vs what was specified on the docs.

I also reordered the options to match the API response order.

Added:
- are_uploads_public
- autosave_progress (docs have this as `autocomplete_progress`)
- free_form_navigation
- hide_navigation
- pro_subdomain_enabled
- show_cookie_consent
- show_key_hint_on_choices
- show_number_of_submissions
- show_question_number
- show_time_to_complete
- use_lead_qualification

Reordered:
- language
- is_public
- progress_bar
- show_progress_bar
- show_typeform_branding